+++ docs/changelog.md
@@ -3,12 +3,13 @@
 ## v0.3.1
 
 - Fix minor bug of the report feature: update TAGS and filters
+- Add external arguments for the report feature: `{{ arg }}` will be replaced by `value` in the report template
 
 ## v0.3.0
 
-- Code refactoring: datastores -> api -> commands (e.g., code under the api folder is based on code under the datastore folder)
+- Code refactoring: datastore -> api -> commands (e.g., code under the `api` folder is based on code under the `datastore` folder)
 - Add Python API: (1) Connection class with create, close, destroy, export, graph, load, and query methods; and (2) create, destroy, export, graph, load, query, and report functions
-- Update CLI of report and query commands: replace --file with --inputfile
+- Update CLI of report and query commands: replace `--file` with `--inputfile`
 - Update the documentation
 
 ## v0.2.1